Fix prior patch, use join_path() instead of concatenation
When testing to see if the protected location is equal to or an
ancestor to mellon_root the paths were formed by concatenating
a path seperator slash, however that might lead to double slashes.
The join_path() utility was writen expressely to handle such situations
and should have been used instead.