Unify command line args between Mellon & OIDC clients
k-h-c-i was originally developed to support Mellon and as such
several command line args and internal variables (especially in the
options namesapece) had Mellon specific names. However many of these
variables can be shared with OIDC. This is a first pass at migrating
"mellon" specific symbols to generic "client" symbols.

The old names can still be used, they are mapped to the new names but
if an old name is used a Deprecation warning is emitted which directs
the user to the new name.