Introduce module files utils.py and keycloak_rest.py
This is the start of a file refactoring effort. The existing files had
grown too large and were not dedicated to a single purpose. With
utilities and the REST interface now split out it's easier to share
functionality and keeps functionality logically grouped
together. Although this covers the bulk of the refactoring there are
still some loose ends to tie up. Mellon and ODIC clients pass sanity
tests as well as the command line tool.