Man page edits, rename descriptor originate method
Rename the client-originate method "descriptor" to "native",
"descriptor" as a name did not reflect the meaning.

Make sure command line args in tool match what is documented in the
man page.

Get rid of final_actions() method, it wasn't doing anything useful.

Fix log message in RESTError constructor, it was logging at the error
level but it should have been logging at the debug level because
RESTError's are caught elsewhere and some are expected and will be
translated into a more appropriate response/action.