Remove stages option
It's not clear that you can easily partition operations
to those involving only the client or only the server. For example the
mellon client needs to contact the keycloak server to get the realm's
IdP metadata, but the realm might not have been created yet and what
does it mean "client only" and yet invoke the server? The stages
option might be useful but it needs to be thought out more. I'm not
sure there is a valid situation where only client, or only server,
actions are run or that we want to support that.